+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix multiple issues in `id_translation.translator_typing`.
+
 ## [1.1.0] - 2026-02-22
 
 ### Added